DDT {MASS}	R Documentation

DDT in Kale

Description

A numeric vector of 15 measurements by different laboratories of the pesticide DDT in kale, in ppm (parts per million) using the multiple pesticide residue measurement.

Usage

DDT

Source

C. E. Finsterwalder (1976) Collaborative study of an extension of the Mills et al method for the determination of pesticide residues in food. J. Off. Anal. Chem. 59, 169–171

R. G. Staudte and S. J. Sheather (1990) Robust Estimation and Testing. Wiley

[Package MASS version 7.2-48 Index]