Nutritional and Marketing Information on US Cereals

The UScereal data frame has 65 rows and 11 columns. The data come from the 1993 ASA Statistical Graphics Exposition, and are taken from the mandatory F&DA food label. The data have been normalized here to a portion of one American cup.

UScereal

Format

This data frame contains the following columns:

mfr

Manufacturer, represented by its first initial: G=General Mills, K=Kelloggs, N=Nabisco, P=Post, Q=Quaker Oats, R=Ralston Purina.

calories

number of calories in one portion.

protein

grams of protein in one portion.

fat

grams of fat in one portion.

sodium

milligrams of sodium in one portion.

fibre

grams of dietary fibre in one portion.

carbo

grams of complex carbohydrates in one portion.

sugars

grams of sugars in one portion.

shelf

display shelf (1, 2, or 3, counting from the floor).

potassium

grams of potassium.

vitamins

vitamins and minerals (none, enriched, or 100%).

Source

The original data are available at http://lib.stat.cmu.edu/datasets/1993.expo/.

References

Venables, W. N. and Ripley, B. D. (1999) Modern Applied Statistics with S-PLUS. Third Edition. Springer.