cabbages {MASS}	R Documentation

Data from a cabbage field trial

Description

The cabbages data set has 60 observations and 4 variables

Usage

cabbages

Format

This data frame contains the following columns:

Cult

Factor giving the cultivar of the cabbage, two levels: c39 and c52.

Date

Factor specifying one of three planting dates: d16, d20 or d21.

HeadWt

Weight of the cabbage head, presumably in kg.

VitC

Ascorbic acid content, in undefined units.

Source

Rawlings, J. O. (1988) Applied Regression Analysis: A Research Tool. Wadsworth and Brooks/Cole. Example 8.4, page 219. (Rawlings cites the original source as the files of the late Dr Gertrude M Cox.)

References

Venables, W. N. and Ripley, B. D. (1999) Modern Applied Statistics with S-PLUS. Third Edition. Springer.

[Package MASS version 7.3-43 Index]