Remission Times of Leukaemia Patients

A data frame from a trial of 42 leukaemia patients. Some were treated with the drug 6-mercaptopurine and the rest are controls. The trial was designed as matched pairs, both withdrawn from the trial when either came out of remission.

gehan

Format

This data frame contains the following columns:

pair

label for pair.

time

remission time in weeks.

cens

censoring, 0/1.

treat

treatment, control or 6-MP.

Source

Cox, D. R. and Oakes, D. (1984) Analysis of Survival Data. Chapman & Hall, p. 7. Taken from

Gehan, E.A. (1965) A generalized Wilcoxon test for comparing arbitrarily single-censored samples. Biometrika 52, 203--233.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

Examples

library(survival)
gehan.surv <- survfit(Surv(time, cens) ~ treat, data = gehan,
     conf.type = "log-log")
summary(gehan.surv)

#> Call: survfit(formula = Surv(time, cens) ~ treat, data = gehan, conf.type = "log-log")
#> 
#>                 treat=6-MP 
#>  time n.risk n.event survival std.err lower 95% CI upper 95% CI
#>     6     21       3    0.857  0.0764        0.620        0.952
#>     7     17       1    0.807  0.0869        0.563        0.923
#>    10     15       1    0.753  0.0963        0.503        0.889
#>    13     12       1    0.690  0.1068        0.432        0.849
#>    16     11       1    0.627  0.1141        0.368        0.805
#>    22      7       1    0.538  0.1282        0.268        0.747
#>    23      6       1    0.448  0.1346        0.188        0.680
#> 
#>                 treat=control 
#>  time n.risk n.event survival std.err lower 95% CI upper 95% CI
#>     1     21       2   0.9048  0.0641      0.67005        0.975
#>     2     19       2   0.8095  0.0857      0.56891        0.924
#>     3     17       1   0.7619  0.0929      0.51939        0.893
#>     4     16       2   0.6667  0.1029      0.42535        0.825
#>     5     14       2   0.5714  0.1080      0.33798        0.749
#>     8     12       4   0.3810  0.1060      0.18307        0.578
#>    11      8       2   0.2857  0.0986      0.11656        0.482
#>    12      6       2   0.1905  0.0857      0.05948        0.377
#>    15      4       1   0.1429  0.0764      0.03566        0.321
#>    17      3       1   0.0952  0.0641      0.01626        0.261
#>    22      2       1   0.0476  0.0465      0.00332        0.197
#>    23      1       1   0.0000     NaN           NA           NA
#> 

survreg(Surv(time, cens) ~ factor(pair) + treat, gehan, dist = "exponential")

#> Call:
#> survreg(formula = Surv(time, cens) ~ factor(pair) + treat, data = gehan, 
#>     dist = "exponential")
#> 
#> Coefficients:
#>    (Intercept)  factor(pair)2  factor(pair)3  factor(pair)4  factor(pair)5 
#>      2.0702861      2.1476909      1.8329493      1.7718527      1.4682566 
#>  factor(pair)6  factor(pair)7  factor(pair)8  factor(pair)9 factor(pair)10 
#>      1.8954775      0.5583010      2.5187140      2.2970513      2.4862208 
#> factor(pair)11 factor(pair)12 factor(pair)13 factor(pair)14 factor(pair)15 
#>      1.0524472      1.8270477      1.6772567      1.7778672      2.0859913 
#> factor(pair)16 factor(pair)17 factor(pair)18 factor(pair)19 factor(pair)20 
#>      3.0634288      0.7996252      1.5855018      1.4083884      0.4023946 
#> factor(pair)21   treatcontrol 
#>      1.9698390     -1.7671562 
#> 
#> Scale fixed at 1 
#> 
#> Loglik(model)= -101.6   Loglik(intercept only)= -116.8
#> 	Chisq= 30.27 on 21 degrees of freedom, p= 0.0866 
#> n= 42 

summary(survreg(Surv(time, cens) ~ treat, gehan, dist = "exponential"))

#> 
#> Call:
#> survreg(formula = Surv(time, cens) ~ treat, data = gehan, dist = "exponential")
#>               Value Std. Error     z       p
#> (Intercept)   3.686      0.333 11.06 < 2e-16
#> treatcontrol -1.527      0.398 -3.83 0.00013
#> 
#> Scale fixed at 1 
#> 
#> Exponential distribution
#> Loglik(model)= -108.5   Loglik(intercept only)= -116.8
#> 	Chisq= 16.49 on 1 degrees of freedom, p= 4.9e-05 
#> Number of Newton-Raphson Iterations: 4 
#> n= 42 
#> 

summary(survreg(Surv(time, cens) ~ treat, gehan))

#> 
#> Call:
#> survreg(formula = Surv(time, cens) ~ treat, data = gehan)
#>               Value Std. Error     z       p
#> (Intercept)   3.516      0.252 13.96 < 2e-16
#> treatcontrol -1.267      0.311 -4.08 4.5e-05
#> Log(scale)   -0.312      0.147 -2.12   0.034
#> 
#> Scale= 0.732 
#> 
#> Weibull distribution
#> Loglik(model)= -106.6   Loglik(intercept only)= -116.4
#> 	Chisq= 19.65 on 1 degrees of freedom, p= 9.3e-06 
#> Number of Newton-Raphson Iterations: 5 
#> n= 42 
#> 

gehan.cox <- coxph(Surv(time, cens) ~ treat, gehan)
summary(gehan.cox)

#> Call:
#> coxph(formula = Surv(time, cens) ~ treat, data = gehan)
#> 
#>   n= 42, number of events= 30 
#> 
#>                coef exp(coef) se(coef)     z Pr(>|z|)    
#> treatcontrol 1.5721    4.8169   0.4124 3.812 0.000138 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#>              exp(coef) exp(-coef) lower .95 upper .95
#> treatcontrol     4.817     0.2076     2.147     10.81
#> 
#> Concordance= 0.69  (se = 0.041 )
#> Likelihood ratio test= 16.35  on 1 df,   p=5e-05
#> Wald test            = 14.53  on 1 df,   p=1e-04
#> Score (logrank) test = 17.25  on 1 df,   p=3e-05
#>