hist.scott {MASS}	R Documentation

Plot a Histogram with Automatic Bin Width Selection

Description

Plot a histogram with automatic bin width selection, using the Scott or Freedman–Diaconis formulae.

Usage


hist.scott(x, prob = TRUE, xlab = deparse(substitute(x)), ...)
hist.FD(x, prob = TRUE, xlab = deparse(substitute(x)), ...)

Arguments

x
A data vector

prob
Should the plot have unit area, so be a density estimate?

xlab, ...
Further arguments to hist.

Value

For the nclass.* functions, the suggested number of classes.

Side Effects

Plot a histogram.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Springer.

See Also

hist

[Package MASS version 7.3-50 Index]