quine {MASS}	R Documentation

Absenteeism from School in Rural New South Wales

Description

The quine data frame has 146 rows and 5 columns. Children from Walgett, New South Wales, Australia, were classified by Culture, Age, Sex and Learner status and the number of days absent from school in a particular school year was recorded.

Usage

quine

Format

This data frame contains the following columns:

Eth

ethnic background: Aboriginal or Not, ("A" or "N").

Sex

sex: factor with levels ("F" or "M").

Age

age group: Primary ("F0"), or forms "F1," "F2" or "F3".

Lrn

learner status: factor with levels Average or Slow learner, ("AL" or "SL").

Days

days absent from school in the year.

Source

S. Quine, quoted in Aitkin, M. (1978) The analysis of unbalanced cross classifications (with discussion). Journal of the Royal Statistical Society series A 141, 195–223.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

[Package MASS version 7.2-48 Index]