rotifer {MASS}	R Documentation

Numbers of Rotifers by Fluid Density

Description

The data give the numbers of rotifers falling out of suspension for different fluid densities. There are two species, pm Polyartha major and kc, Keratella cochlearis and for each species the number falling out and the total number are given.

Usage

rotifer

Format

density

specific density of fluid.

pm.y

number falling out for P. major.

pm.total

total number of P. major.

kc.y

number falling out for K. cochlearis.

kc.tot

total number of K. cochlearis.

Source

D. Collett (1991) Modelling Binary Data. Chapman & Hall. p. 217

[Package MASS version 7.2-48 Index]