summary.loglm {MASS}	R Documentation

Summary Method Function for Objects of Class 'loglm'

Description

Returns a summary list for log-linear models fitted by iterative proportional scaling using loglm.

Usage


## S3 method for class 'loglm'
summary(object, fitted = FALSE, ...)

Arguments

object
a fitted loglm model object.

fitted
if TRUE return observed and expected frequencies in the result. Using fitted = TRUE may necessitate re-fitting the object.

...
arguments to be passed to or from other methods.

Details

This function is a method for the generic function summary() for class "loglm". It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.loglm(x) regardless of the class of the object.

Value

a list is returned for use by print.summary.loglm. This has components

formula
the formula used to produce object

tests
the table of test statistics (likelihood ratio, Pearson) for the fit.

oe
if fitted = TRUE, an array of the observed and expected frequencies, otherwise NULL.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

loglm, summary

[Package MASS version 7.3-50 Index]