summary.negbin {MASS}	R Documentation

Summary Method Function for Objects of Class 'negbin'

Description

Identical to summary.glm, but with three lines of additional output: the ML estimate of theta, its standard error, and twice the log-likelihood function.

Usage


## S3 method for class 'negbin'
summary(object, dispersion = 1, correlation = FALSE, ...)

Arguments

object
fitted model object of class negbin inheriting from glm and lm. Typically the output of glm.nb.

dispersion
as for summary.glm, with a default of 1.

correlation
as for summary.glm.

...
arguments passed to or from other methods.

Details

summary.glm is used to produce the majority of the output and supply the result. This function is a method for the generic function summary() for class "negbin". It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.negbin(x) regardless of the class of the object.

Value

As for summary.glm; the additional lines of output are not included in the resultant object.

Side Effects

A summary table is produced as for summary.glm, with the additional information described above.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

summary, glm.nb, negative.binomial, anova.negbin

Examples


summary(glm.nb(Days ~ Eth*Age*Lrn*Sex, quine, link = log))

[Package MASS version 7.3-50 Index]