survey {MASS}	R Documentation

Student Survey Data

Description

This data frame contains the responses of 237 Statistics I students at the University of Adelaide to a number of questions.

Usage

survey

Format

The components of the data frame are:

Sex

The sex of the student. (Factor with levels "Male" and "Female".)

Wr.Hnd

span (distance from tip of thumb to tip of little finger of spread hand) of writing hand, in centimetres.

NW.Hnd

span of non-writing hand.

W.Hnd

writing hand of student. (Factor, with levels "Left" and "Right".)

Fold

“Fold your arms! Which is on top” (Factor, with levels "R on L", "L on R", "Neither".)

Pulse

pulse rate of student (beats per minute).

Clap

‘Clap your hands! Which hand is on top?’ (Factor, with levels "Right", "Left", "Neither".)

Exer

how often the student exercises. (Factor, with levels "Freq" (frequently), "Some", "None".)

Smoke

how much the student smokes. (Factor, levels "Heavy", "Regul" (regularly), "Occas" (occasionally), "Never".)

Height

height of the student in centimetres.

M.I

whether the student expressed height in imperial (feet/inches) or metric (centimetres/metres) units. (Factor, levels "Metric", "Imperial".)

Age

age of the student in years.

References

Venables, W. N. and Ripley, B. D. (1999) Modern Applied Statistics with S-PLUS. Third Edition. Springer.

[Package MASS version 7.2-48 Index]