MASS-internal {MASS}	R Documentation

Internal MASS functions

Description

Internal MASS functions.

Usage

enlist(vec)
fbeta(x, alpha, beta)
frequency.polygon(x, nclass = nclass.freq(x), xlab="", ylab="", ...)
nclass.freq(x)
neg.bin(theta = stop("'theta' must be given"))
negexp.SSival(mCall, data, LHS)

Details

These are not intended to be called by the user. Some are for compatibility with earlier versions of MASS (the book).

[Package MASS version 7.2-48 Index]