con2tr {MASS}	R Documentation

Convert Lists to Data Frames for use by Lattice

Description

Convert lists to data frames for use by lattice.

Usage

con2tr(obj)

Arguments

obj 	 A list of components x, y and z as passed to contour.

Details

con2tr repeats the x and y components suitably to match the vector z.

Value

A data frame suitable for passing to lattice (formerly trellis) functions.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

[Package MASS version 7.2-48 Index]