glmmPQL {MASS}	R Documentation

Fit Generalized Linear Mixed Models via PQL

Description

Fit a GLMM model with multivariate normal random effects, using Penalized Quasi-Likelihood.

Usage

glmmPQL(fixed, random, family, data, correlation, weights,
        control, niter = 10, verbose = TRUE, ...)

Arguments

fixed 	 a two-sided linear formula giving fixed-effects part of the model.

random 	 a formula or list of formulae describing the random effects.

family 	 a GLM family.

data 	 an optional data frame used as the first place to find variables in the formulae.

correlation 	 an optional correlation structure.

weights 	 optional case weights as in glm.

control 	 an optional argument to be passed to lme.

niter 	 maximum number of iterations.

verbose 	 logical: print out record of iterations?

... 	 Further arguments for lme.

Details

glmmPQL works by repeated calls to lme, so package nlme will be loaded at first use if necessary.

Value

A object of class "lme": see lmeObject.

References

Schall, R. (1991) Estimation in generalized linear models with random effects. Biometrika 78, 719–727.

Breslow, N. E. and Clayton, D. G. (1993) Approximate inference in generalized linear mixed models. Journal of the American Statistical Association 88, 9–25.

Wolfinger, R. and O'Connell, M. (1993) Generalized linear mixed models: a pseudo-likelihood approach. Journal of Statistical Computation and Simulation 48, 233–243.

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

lme

Examples

library(nlme) # will be loaded automatically if omitted
summary(glmmPQL(y ~ trt + I(week > 2), random = ~ 1 | ID,
                family = binomial, data = bacteria))

[Package MASS version 7.2-48 Index]