ldahist {MASS}	R Documentation

Histograms or Density Plots of Multiple Groups

Description

Plot histograms or density plots of data on a single Fisher linear discriminant.

Usage

ldahist(data, g, nbins = 25, h, x0 = - h/1000, breaks,
        xlim = range(breaks), ymax = 0, width,
        type = c("histogram", "density", "both"),
        sep = (type != "density"),
        col = 5, xlab = deparse(substitute(data)), bty = "n", ...)

Arguments

data 	 vector of data. Missing values (NAs) are allowed and omitted.

g 	 factor or vector giving groups, of the same length as data.

nbins 	 Suggested number of bins to cover the whole range of the data.

h 	 The bin width (takes precedence over nbins).

x0 	 Shift for the bins - the breaks are at x0 + h * (..., -1, 0, 1, ...)

breaks 	 The set of breakpoints to be used. (Usually omitted, takes precedence over h and nbins).

xlim 	 The limits for the x-axis.

ymax 	 The upper limit for the y-axis.

width 	 Bandwidth for density estimates. If missing, the Sheather-Jones selector is used for each group separately.

type 	 Type of plot.

sep 	 Whether there is a separate plot for each group, or one combined plot.

col 	 The colour number for the bar fill.

xlab 	 label for the plot x-axis. By default, this will be the name of data.

bty 	 The box type for the plot - defaults to none.

... 	 additional arguments to polygon.

Side Effects

Histogram and/or density plots are plotted on the current device.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

plot.lda.

[Package MASS version 7.2-48 Index]