mca {MASS}	R Documentation

Multiple Correspondence Analysis

Description

Computes a multiple correspondence analysis of a set of factors.

Usage

mca(df, nf = 2, abbrev = FALSE)

Arguments

df 	 A data frame containing only factors

nf 	 The number of dimensions for the MCA. Rarely 3 might be useful.

abbrev 	 Should the vertex names be abbreviated? By default these are of the form ‘factor.level’ but if abbrev = TRUE they are just ‘level’ which will suffice if the factors have distinct levels.

Value

An object of class "mca", with components

rs 	 The coordinates of the rows, in nf dimensions.

cs 	 The coordinates of the column vertices, one for each level of each factor.

fs 	 Weights for each row, used to interpolate additional factors in predict.mca.

p 	 The number of factors

d 	 The singular values for the nf dimensions.

call 	 The matched call.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

predict.mca, plot.mca, corresp

Examples

farms.mca <- mca(farms, abbrev=TRUE)
farms.mca
plot(farms.mca)

[Package MASS version 7.2-48 Index]