pairs.lda {MASS}	R Documentation

Produce Pairwise Scatterplots from an 'lda' Fit

Description

Pairwise scatterplot of the data on the linear discriminants.

Usage

## S3 method for class 'lda':
pairs(x, labels = colnames(x), panel = panel.lda,
     dimen, abbrev = FALSE, ..., cex=0.7, type = c("std", "trellis"))

Arguments

x 	 Object of class "lda".

labels 	 vector of character strings for labelling the variables.

panel 	 panel function to plot the data in each panel.

dimen 	 The number of linear discriminants to be used for the plot; if this exceeds the number determined by x the smaller value is used.

abbrev 	 whether the group labels are abbreviated on the plots. If abbrev > 0 this gives minlength in the call to abbreviate.

... 	 additional arguments for pairs.default.

cex 	 graphics parameter cex for labels on plots.

type 	 type of plot. The default is in the style of pairs.default; the style "trellis" uses the Trellis function splom.

Details

This function is a method for the generic function pairs() for class "lda". It can be invoked by calling pairs(x) for an object x of the appropriate class, or directly by calling pairs.lda(x) regardless of the class of the object.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

pairs

[Package MASS version 7.2-48 Index]