plot.mca {MASS}	R Documentation

Plot Method for Objects of Class 'mca'

Description

Plot a multiple correspondence analysis.

Usage

## S3 method for class 'mca':
plot(x, rows = TRUE, col, cex = par("cex"), ...)

Arguments

x 	 An object of class "mca".

rows 	 Should the coordinates for the rows be plotted, or just the vertices for the levels?

col, cex 	 The colours and cex to be used for the row points and level vertices respectively.

... 	 Additional parameters to plot.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

mca, predict.mca

Examples

plot(mca(farms, abbrev = TRUE))

[Package MASS version 7.2-48 Index]