predict.glmmPQL {MASS}	R Documentation

Predict Method for glmmPQL Fits

Description

Obtains predictions from a fitted generalized linear model with random effects.

Usage

## S3 method for class 'glmmPQL':
predict(object, newdata = NULL, type = c("link", "response"),
       level, na.action = na.pass, ...)

Arguments

object 	 a fitted object of class inheriting from "glmmPQL".

newdata 	 optionally, a data frame in which to look for variables with which to predict.

type 	 the type of prediction required. The default is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable. Thus for a default binomial model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities.

level 	 an optional integer vector giving the level(s) of grouping to be used in obtaining the predictions. Level values increase from outermost to innermost grouping, with level zero corresponding to the population predictions. Defaults to the highest or innermost level of grouping.

na.action 	 function determining what should be done with missing values in newdata. The default is to predict NA.

... 	 further arguments passed to or from other methods.

Value

If level is a single integer, a vector otherwise a data frame.

See Also

glmmPQL, predict.lme.

Examples

fit <- glmmPQL(y ~ trt + I(week > 2), random = ~1 |  ID,
               family = binomial, data = bacteria)
predict(fit, bacteria, level = 0, type="response")
predict(fit, bacteria, level = 1, type="response")

[Package MASS version 7.2-48 Index]