predict.mca {MASS}	R Documentation

Predict Method for Class 'mca'

Description

Used to compute coordinates for additional rows or additional factors in a multiple correspondence analysis.

Usage

## S3 method for class 'mca':
predict(object, newdata, type = c("row", "factor"), ...)

Arguments

object 	 An object of class "mca", usually the result of a call to mca.

newdata 	 A data frame containing either additional rows of the factors used to fit object or additional factors for the cases used in the original fit.

type 	 Are predictions required for further rows or for new factors?

... 	 Additional arguments from predict: unused.

Value

If type = "row", the coordinates for the additional rows.
If type = "factor", the coordinates of the column vertices for the levels of the new factors.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See Also

mca, plot.mca

[Package MASS version 7.2-48 Index]