Relative Curvature Measures for Non-Linear Regression

Calculates the root mean square parameter effects and intrinsic relative curvatures, \(c^\theta\) and \(c^\iota\), for a fitted nonlinear regression, as defined in Bates & Watts, section 7.3, p. 253ff

rms.curv(obj)

Arguments

obj
Fitted model object of class "nls". The model must be fitted using the default algorithm.

Value

A list of class rms.curv with components pc and ic for parameter effects and intrinsic relative curvatures multiplied by sqrt(F), ct and ci for \(c^\theta\) and \(c^\iota\) (unmultiplied), and C the C-array as used in section 7.3.1 of Bates & Watts.

Details

The method of section 7.3.1 of Bates & Watts is implemented. The function deriv3 should be used generate a model function with first derivative (gradient) matrix and second derivative (Hessian) array attributes. This function should then be used to fit the nonlinear regression model.

A print method, print.rms.curv, prints the pc and ic components only, suitably annotated.

If either pc or ic exceeds some threshold (0.3 has been suggested) the curvature is unacceptably high for the planar assumption.

References

Bates, D. M, and Watts, D. G. (1988) Nonlinear Regression Analysis and its Applications. Wiley, New York.

See also

Examples

# The treated sample from the Puromycin data
mmcurve <- deriv3(~ Vm * conc/(K + conc), c("Vm", "K"),
                  function(Vm, K, conc) NULL)
Treated <- Puromycin[Puromycin$state == "treated", ]
(Purfit1 <- nls(rate ~ mmcurve(Vm, K, conc), data = Treated,
                start = list(Vm=200, K=0.1)))

#> Nonlinear regression model
#>   model: rate ~ mmcurve(Vm, K, conc)
#>    data: Treated
#>        Vm         K 
#> 212.68363   0.06412 
#>  residual sum-of-squares: 1195
#> 
#> Number of iterations to convergence: 6 
#> Achieved convergence tolerance: 6.096e-06

rms.curv(Purfit1)

#> Parameter effects: c^theta x sqrt(F) = 0.2121 
#>         Intrinsic: c^iota  x sqrt(F) = 0.092 

##Parameter effects: c^theta x sqrt(F) = 0.2121
##        Intrinsic: c^iota  x sqrt(F) = 0.092