Summary Method Function for Objects of Class 'negbin'

Identical to summary.glm, but with three lines of additional output: the ML estimate of theta, its standard error, and twice the log-likelihood function.

# S3 method for negbin
summary(object, dispersion = 1, correlation = FALSE, ...)

Arguments

object
fitted model object of class negbin inheriting from glm and lm. Typically the output of glm.nb.

dispersion
as for summary.glm, with a default of 1.

correlation
as for summary.glm.

...
arguments passed to or from other methods.

Value

As for summary.glm; the additional lines of output are not included in the resultant object.

Side Effects

A summary table is produced as for summary.glm, with the additional information described above.

Details

summary.glm is used to produce the majority of the output and supply the result. This function is a method for the generic function summary() for class "negbin". It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.negbin(x) regardless of the class of the object.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See also

Examples

summary(glm.nb(Days ~ Eth*Age*Lrn*Sex, quine, link = log))

#> 
#> Call:
#> glm.nb(formula = Days ~ Eth * Age * Lrn * Sex, data = quine, 
#>     link = log, init.theta = 1.928360145)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -3.2377  -0.9079  -0.2019   0.5173   1.7043  
#> 
#> Coefficients: (4 not defined because of singularities)
#>                       Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)             3.0564     0.3760   8.128 4.38e-16 ***
#> EthN                   -0.1386     0.5334  -0.260 0.795023    
#> AgeF1                  -0.6227     0.5125  -1.215 0.224334    
#> AgeF2                  -2.3632     1.0770  -2.194 0.028221 *  
#> AgeF3                  -0.3784     0.4546  -0.832 0.405215    
#> LrnSL                  -1.9577     0.9967  -1.964 0.049493 *  
#> SexM                   -0.4914     0.5104  -0.963 0.335653    
#> EthN:AgeF1              0.1029     0.7123   0.144 0.885175    
#> EthN:AgeF2             -0.5546     1.6798  -0.330 0.741297    
#> EthN:AgeF3              0.0633     0.6396   0.099 0.921159    
#> EthN:LrnSL              2.2588     1.3019   1.735 0.082743 .  
#> AgeF1:LrnSL             2.6421     1.0821   2.442 0.014618 *  
#> AgeF2:LrnSL             4.8585     1.4423   3.369 0.000755 ***
#> AgeF3:LrnSL                 NA         NA      NA       NA    
#> EthN:SexM              -0.7524     0.7220  -1.042 0.297400    
#> AgeF1:SexM              0.4092     0.8299   0.493 0.621973    
#> AgeF2:SexM              3.1098     1.1655   2.668 0.007624 ** 
#> AgeF3:SexM              1.1145     0.6365   1.751 0.079926 .  
#> LrnSL:SexM              1.5900     1.1499   1.383 0.166750    
#> EthN:AgeF1:LrnSL       -3.5493     1.4270  -2.487 0.012876 *  
#> EthN:AgeF2:LrnSL       -3.3315     2.0919  -1.593 0.111256    
#> EthN:AgeF3:LrnSL            NA         NA      NA       NA    
#> EthN:AgeF1:SexM        -0.3105     1.2055  -0.258 0.796735    
#> EthN:AgeF2:SexM         0.3469     1.7965   0.193 0.846875    
#> EthN:AgeF3:SexM         0.8329     0.8970   0.929 0.353092    
#> EthN:LrnSL:SexM        -0.1639     1.5250  -0.107 0.914411    
#> AgeF1:LrnSL:SexM       -2.4285     1.4201  -1.710 0.087246 .  
#> AgeF2:LrnSL:SexM       -4.1914     1.6201  -2.587 0.009679 ** 
#> AgeF3:LrnSL:SexM            NA         NA      NA       NA    
#> EthN:AgeF1:LrnSL:SexM   2.1711     1.9192   1.131 0.257963    
#> EthN:AgeF2:LrnSL:SexM   2.1029     2.3444   0.897 0.369718    
#> EthN:AgeF3:LrnSL:SexM       NA         NA      NA       NA    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for Negative Binomial(1.9284) family taken to be 1)
#> 
#>     Null deviance: 272.29  on 145  degrees of freedom
#> Residual deviance: 167.45  on 118  degrees of freedom
#> AIC: 1097.3
#> 
#> Number of Fisher Scoring iterations: 1
#> 
#> 
#>               Theta:  1.928 
#>           Std. Err.:  0.269 
#> 
#>  2 x log-likelihood:  -1039.324