Summary Method for Robust Linear Models

summary method for objects of class "rlm"

# S3 method for rlm
summary(object, method = c("XtX", "XtWX"), correlation = FALSE, ...)

Arguments

object
the fitted model. This is assumed to be the result of some fit that produces an object inheriting from the class rlm, in the sense that the components returned by the rlm function will be available.

method
Should the weighted (by the IWLS weights) or unweighted cross-products matrix be used?

correlation
logical. Should correlations be computed (and printed)?

...
arguments passed to or from other methods.

Value

If printing takes place, only a null value is returned. Otherwise, a list is returned with the following components. Printing always takes place if this function is invoked automatically as a method for the summary function.

correlation

The computed correlation coefficient matrix for the coefficients in the model.

cov.unscaled

The unscaled covariance matrix; i.e, a matrix such that multiplying it by an estimate of the error variance produces an estimated covariance matrix for the coefficients.

sigma

The scale estimate.

stddev

A scale estimate used for the standard errors.

df

The number of degrees of freedom for the model and for residuals.

coefficients

A matrix with three columns, containing the coefficients, their standard errors and the corresponding t statistic.

terms

The terms object used in fitting this model.

Details

This function is a method for the generic function summary() for class "rlm". It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.rlm(x) regardless of the class of the object.

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.

See also

Examples

summary(rlm(calls ~ year, data = phones, maxit = 50))

#> 
#> Call: rlm(formula = calls ~ year, data = phones, maxit = 50)
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -18.314  -5.953  -1.681  26.460 173.769 
#> 
#> Coefficients:
#>             Value     Std. Error t value  
#> (Intercept) -102.6222   26.6082    -3.8568
#> year           2.0414    0.4299     4.7480
#> 
#> Residual standard error: 9.032 on 22 degrees of freedom